split einsum global tests (#7713)
* split einsum global tests

* auto format by CI
